FFTA.jl

Documentation for FFTA.jl

FFTA.fft — Function

Perform a fast Fourier transform of a vector. Preserves types given by the user.

Arguments

x::AbstractVector: The vector to transform.

Examples

julia> x = rand(ComplexF64, 10)

julia> y = fft(x)

Perform a fast Fourier transform of a matrix. Preserves types given by the user.

Arguments

x::AbstractMatrix: The matrix to transform (columnwise then rowwise).

Examples

julia> x = rand(ComplexF64, 10, 10)

julia> y = fft(x)

FFTA.bfft — Function

bfft(x::AbstractArray{T, 1}) -> Any

Perform a backward fast Fourier transform of a vector, where "backward" indicates the same output signal down to a constant factor. Preserves types given by the user.

Arguments

x::AbstractVector: The vector to transform

Examples

julia> x = rand(ComplexF64, 10)

julia> y = bfft(x)

julia> z = fft(y)

julia> x ≈ z/10
true

Perform a backward fast Fourier transform of a matrix, where "backward" indicates the same output signal down to a constant factor. Preserves types given by the user.

Arguments

x::AbstractMatrix: The matrix to transform

Examples

julia> x = rand(ComplexF64, 10, 10)

julia> y = bfft(x)

julia> z = fft(y)

julia> x ≈ z/100
true