Development Documentation

Here is the documentation for key parts of the development side of the package.

FFTA.CallGraphNode — Type

struct CallGraphNode{T}

Node of a call graph

Arguments

left: Offset to the left child node

right: Offset to the right child node

type: Object representing the type of FFT

sz: Size of this FFT

FFTA.CallGraph — Type

struct CallGraph{T<:Complex}

Object representing a graph of FFT Calls

Arguments

nodes: Nodes keeping track of the graph

workspace: Preallocated Workspace

BLUESTEIN_CUTOFF: Minimum prime that will be FFTed with the Bluestein algorithm, below which the O(N^2) DFT is used.

FFTA.CallGraphNode! — Function

CallGraphNode!(
    nodes::Array{FFTA.CallGraphNode{T}, 1},
    N::Int64,
    workspace::Array{Array{T, 1}, 1},
    BLUESTEIN_CUTOFF::Int64,
    s_in::Int64,
    s_out::Int64
) -> Int64

Recursively instantiate a set of CallGraphNodes

Arguments

nodes: A vector (which gets expanded) of CallGraphNodes

N: The size of the FFT

workspace: A vector (which gets expanded) of preallocated workspaces

s_in: The stride of the input

s_out: The stride of the output

Missing docstring.

Missing docstring for fft!. Check Documenter's build log for details.

FFTA.fft_dft! — Function

fft_dft!(
    out::AbstractArray{T, 1},
    in::AbstractArray{T, 1},
    N::Int64,
    start_out::Int64,
    stride_out::Int64,
    start_in::Int64,
    stride_in::Int64,
    w
)

Discrete Fourier Transform, O(N^2) algorithm, in place.

Arguments

out: Output vector

in: Input vector

N: Size of the transform

start_out: Index of the first element of the output vector

stride_out: Stride of the output vector

start_in: Index of the first element of the input vector

stride_in: Stride of the input vector

w: The value cispi(direction_sign(d) * 2 / N)

FFTA.fft_pow3! — Function

fft_pow3!(
    out::AbstractArray{T, 1},
    in::AbstractArray{U, 1},
    N::Int64,
    start_out::Int64,
    stride_out::Int64,
    start_in::Int64,
    stride_in::Int64,
    w,
    minus120
)

Power of 3 FFT, in place

Arguments

out: Output vector

in: Input vector

N: Size of the transform

start_out: Index of the first element of the output vector

stride_out: Stride of the output vector

start_in: Index of the first element of the input vector

stride_in: Stride of the input vector

w: The value cispi(direction_sign(d) * 2 / N)

plus120: Depending on direction, perform either ±120° rotation

minus120: Depending on direction, perform either ∓120° rotation